move days analyst to new file
